Working application.
Gets a list of all processes.
Starts XBMC on the default path.
Can find if the process XBMC exists.

By starting the XBMC it will bring it upp to the foreground if it is
already started but I will change so it checks if it is already stared
and then bring it up to the foreground manually.